interface_requirement.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aiohttp=3.8.6=pypi_0
aiosignal=1.3.1=pypi_0
altair=5.1.2=pypi_0
annotated-types=0.6.0=pypi_0
anyio=3.7.1=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asttokens=2.4.1=pypi_0
async-lru=2.0.4=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.1.0=pypi_0
babel=2.13.1=pypi_0
beautifulsoup4=4.12.2=pypi_0
bleach=6.1.0=pypi_0
blinker=1.7.0=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.08.22=h06a4308_0
cachetools=5.3.2=pypi_0
certifi=2023.7.22=pypi_0
cffi=1.16.0=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
comm=0.2.0=pypi_0
dataclasses-json=0.6.1=pypi_0
debugpy=1.8.0=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
distro=1.8.0=pypi_0
exceptiongroup=1.1.3=pypi_0
executing=2.0.1=pypi_0
fastjsonschema=2.18.1=pypi_0
filelock=3.13.1=pypi_0
fqdn=1.5.1=pypi_0
frozenlist=1.4.0=pypi_0
fsspec=2023.10.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.40=pypi_0
greenlet=3.0.1=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.1=pypi_0
httpx=0.25.1=pypi_0
huggingface-hub=0.17.3=pypi_0
idna=3.4=pypi_0
importlib-metadata=6.8.0=pypi_0
ipykernel=6.26.0=pypi_0
ipython=8.17.2=pypi_0
ipywidgets=8.1.1=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.2=pypi_0
json5=0.9.14=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=2.4=pypi_0
jsonschema=4.19.2=pypi_0
jsonschema-specifications=2023.7.1=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=8.6.0=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=5.5.0=pypi_0
jupyter-events=0.9.0=pypi_0
jupyter-lsp=2.2.0=pypi_0
jupyter-server=2.10.0=pypi_0
jupyter-server-terminals=0.4.4=pypi_0
jupyterlab=4.0.8=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-server=2.25.0=pypi_0
jupyterlab-widgets=3.0.9=pypi_0
langchain=0.0.331=pypi_0
langsmith=0.0.56=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=pypi_0
marshmallow=3.20.1=pypi_0
matplotlib-inline=0.1.6=pypi_0
mdurl=0.1.2=pypi_0
mistune=3.0.2=pypi_0
multidict=6.0.4=pypi_0
mypy-extensions=1.0.0=pypi_0
mysql-connector=2.2.9=pypi_0
mysql-connector-python=8.2.0=pypi_0
nbclient=0.8.0=pypi_0
nbconvert=7.11.0=pypi_0
nbformat=5.9.2=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.8=pypi_0
notebook=7.0.6=pypi_0
notebook-shim=0.2.3=pypi_0
npm=0.1.1=pypi_0
npx=0.1.1=pypi_0
numpy=1.26.1=pypi_0
openai=0.28.1=pypi_0
openssl=3.0.12=h7f8727e_0
optional-django=0.1.0=pypi_0
overrides=7.4.0=pypi_0
packaging=23.2=pypi_0
pandas=2.1.2=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
passlib=1.7.4=pypi_0
pexpect=4.8.0=pypi_0
pillow=10.1.0=pypi_0
pip=23.3=py310h06a4308_0
platformdirs=3.11.0=pypi_0
prometheus-client=0.18.0=pypi_0
prompt-toolkit=3.0.39=pypi_0
protobuf=4.21.12=pypi_0
psutil=5.9.6=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyarrow=14.0.0=pypi_0
pycparser=2.21=pypi_0
pydantic=2.4.2=pypi_0
pydantic-core=2.10.1=pypi_0
pydeck=0.8.1b0=pypi_0
pygments=2.16.1=pypi_0
python=3.10.13=h955ad1f_0
python-dateutil=2.8.2=pypi_0
python-json-logger=2.0.7=pypi_0
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.1=pypi_0
qtconsole=5.5.0=pypi_0
qtpy=2.4.1=pypi_0
readline=8.2=h5eee18b_0
referencing=0.30.2=pypi_0
regex=2023.10.3=pypi_0
requests=2.31.0=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rich=13.6.0=pypi_0
rpds-py=0.12.0=pypi_0
safetensors=0.4.0=pypi_0
send2trash=1.8.2=pypi_0
setuptools=68.0.0=py310h06a4308_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
sniffio=1.3.0=pypi_0
soupsieve=2.5=pypi_0
sqlalchemy=2.0.23=pypi_0
sqlite=3.41.2=h5eee18b_0
stack-data=0.6.3=pypi_0
streamlit=1.28.1=pypi_0
tenacity=8.2.3=pypi_0
terminado=0.17.1=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.14.1=pypi_0
toml=0.10.2=pypi_0
tomli=2.0.1=pypi_0
toolz=0.12.0=pypi_0
tornado=6.3.3=pypi_0
tqdm=4.66.1=pypi_0
traitlets=5.13.0=pypi_0
transformers=4.35.0=pypi_0
types-python-dateutil=2.8.19.14=pypi_0
typing-extensions=4.8.0=pypi_0
typing-inspect=0.9.0=pypi_0
tzdata=2023.3=pypi_0
tzlocal=5.2=pypi_0
uri-template=1.3.0=pypi_0
urllib3=2.0.7=pypi_0
validators=0.22.0=pypi_0
watchdog=3.0.0=pypi_0
wcwidth=0.2.9=pypi_0
webcolors=1.13=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.6.4=pypi_0
wheel=0.41.2=py310h06a4308_0
widgetsnbextension=4.0.9=pypi_0
xz=5.4.2=h5eee18b_0
yarl=1.9.2=pypi_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h5eee18b_0
